Reorganized src/lib/api/schemas/ files.
