+++ src/lib/db/seed/seed.ts
@@ -10,7 +10,7 @@ import {
 	skills,
 } from "@/lib/db/schema"
 import { resume } from "@/lib/db/seed/resume"
-import { resumeSchema } from "@/lib/db/seed/resume.schema"
+import { resumeSchema } from "@/lib/api/schemas/resume/contract"
 
 // Environment variables.
 config({ path: ".env.local" })
